% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{bind.color.matrixes}
\alias{bind.color.matrixes}
\title{Rbind or cbind color matrixes
 Rbind or cbind color matrixes and optinally add a neutral color in between}
\usage{
bind.color.matrixes(
  my.mat1,
  my.mat2,
  c_or_r = "c",
  add.color.space = 0,
  add.color = "white"
)
}
\arguments{
\item{my.mat1, }{a color matrix used for heatmaps}

\item{my.mat2, }{a color matrix used for heatmaps}

\item{c_or_r, }{shoud be "c" (cbind) or "r" (rbinds)}

\item{add.color.space, }{define how many rows/columns if to add a "neutral" color between the matrix}

\item{add.color, }{what color to add}
}
\value{
a matrix
}
\description{
Rbind or cbind color matrixes
 Rbind or cbind color matrixes and optinally add a neutral color in between
}
