% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Shiny.R
\name{boxplot_SHINY_annotation}
\alias{boxplot_SHINY_annotation}
\title{generic shiny boxplots given an expression set (defaults to pan_gdc) or all three of fdata, pdata and a matrix}
\usage{
boxplot_SHINY_annotation(
  eset = NULL,
  mat = NULL,
  fdata = NULL,
  pdata = NULL,
  color.key = dlfoo2::color_subtypes,
  fdata.altName.column = NULL,
  default.pdata.column = NULL
)
}
\arguments{
\item{mat}{gene identifier}

\item{pdata}{sample data frame with sample annotations}

\item{color.key}{named character vector with sample bxp colors}

\item{fdata.altName.column}{alternative name column, e.g. hgnc_symbol}

\item{default.pdata.column}{name of preferred pdata annotations column (not needed)}

\item{...}{additional parameters}
}
\value{
a shiny app plot will open
}
\description{
Shiny app for plotting annotations (pdata) for genes expression selectable from rows in a matrix (fdata)
}
\seealso{
Other shiny: 
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{shinyPlot_annotation}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other ggplot: 
\code{\link{label_wrap_mod}()}
}
\concept{boxplot}
\concept{ggplot}
\concept{shiny}
\concept{violin}
