% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{chip_genePeaks2occupancyTab}
\alias{chip_genePeaks2occupancyTab}
\title{From a gene/promoter-centered gRanges object - get 1) the number of peaks per promoter and 2) the max peak value from all peaks within one gene promoter}
\usage{
chip_genePeaks2occupancyTab(
  peak2gene.gr,
  value.column = NULL,
  refgene.ref.column = "GENEID",
  refgene.cols = c("GENEID", "SYMBOL"),
  refgene = NULL
)
}
\arguments{
\item{peak2gene.gr, }{a bed-style genomicRanges object with peak positions from chip_peak2gene function}

\item{value.column, }{names of chip peak value column}

\item{refgene.ref.column, }{what refgene column to match on}

\item{refgene.cols, }{what refgene columns to output (condense)}

\item{refgene, }{refgene, what local refgene gRanges object to use}
}
\value{
a data frame with one row per queried refgene entry
}
\description{
From a gene/promoter-centered gRanges object - get 1) the number of peaks per promoter and 2) the max peak value from all peaks within one gene promoter
}
\seealso{
Other old functions: 
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other genomics: 
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other granges: 
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other chip: 
\code{\link{chip_peak2gene_maxValue}()}
}
\concept{chip}
\concept{genomics}
\concept{granges}
\concept{old functions}
