% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{chip_peak2gene}
\alias{chip_peak2gene}
\title{Extract chip calls in the promoter of genes
Wrapper using GenomicRanges::promoters to extract chip peaks from a genomic ranges object located in gene promoters (obtained from UCSC refgene object).}
\usage{
chip_peak2gene(
  my.bed = NULL,
  upstream = 2000,
  downstream = 200,
  refgene = NULL
)
}
\arguments{
\item{my.bed}{a bed-style genomicRanges object with peak positions}

\item{upstream, number}{of bases upstream TSS}

\item{downstream, }{.. of TSS}

\item{refgene, }{what local refgene gRanges object to use}
}
\value{
a genomic ranges object of the chip peaks contained in gene promoters
}
\description{
Extract chip calls in the promoter of genes
Wrapper using GenomicRanges::promoters to extract chip peaks from a genomic ranges object located in gene promoters (obtained from UCSC refgene object).
}
\seealso{
Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other bed: 
\code{\link{chip_peak2gene_maxValue}()}
}
\concept{bed}
\concept{genomics}
\concept{granges}
\concept{old functions}
