% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{chordPlot_ccHCA_to_reTax}
\alias{chordPlot_ccHCA_to_reTax}
\title{chordPlot_ccHCA_to_reTax}
\usage{
chordPlot_ccHCA_to_reTax(
  plot.name = "temp",
  plot.dir = "~/PROJECTS/RCC_ccpRCC_2019/reTaxonomy/Figures/",
 
    cc.dir = "~/PROJECTS/RCC_ccpRCC_2019/ConsensusClust/ConsensusCluster_rcc_t_me_589set_2065genes_reps.1000_maxK.25_innerLinkage.ward_sd.0.01/",
  cc.k = 25,
  group2_levels = c("outlier", "metanephric", "mtscRCC", "dCIMP", "mesRCC", "ccpRCC",
    "chONC", "pONC"),
  group_names = c("cc", "retax"),
  scale_factors = c(1, 3),
  color_key = NULL,
  dend.rev.clusters = NULL,
  dend.color.key = NULL
)
}
\arguments{
\item{plot.name}{name of file to plot}

\item{plot.dir}{directory to plot in}

\item{cc.dir}{directory in which consensus clust results are - parsed using dlfoo2::ConsensusCluster_process}

\item{cc.k}{number of consensusclusters (cannot be higher than performed in cc.dir)}

\item{group2_levels}{factor levels for the re-taxonomy. specifies order of groups in plot}

\item{group_names}{Title of plot}

\item{scale_factors}{what color key to use for annotation. Detaults to dlfoo2::color_subtypes (but could be set to e.g. dlfoo2::color_annotations)}

\item{color_key}{what color key for subgroups.}

\item{dend.rev.clusters}{if to flip clusters.}

\item{dend.color.key}{if a specific color key is to be used for dendogram}

\item{do.gradient}{set to FALSE if to manually override numeric check for gradient or manual fil}
}
\value{
plot pdf to file
}
\description{
specialized function alluvial chordPlot re-taxonomy w consensus HCA
}
\seealso{
Other chordplot: 
\code{\link{circosPlot_cc_hca_tree}()}

Other PlotWrap: 
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other circlize: 
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{links2sectors}()},
\code{\link{rectMatrix}()}
}
\concept{PlotWrap}
\concept{chordplot}
\concept{circlize}
