% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{circosPlot_cc_hca_tree}
\alias{circosPlot_cc_hca_tree}
\title{chordPlot_ccHCA_to_reTax - Plot HCA tree for Consensus Clustering
Here one can shoose to plot with Chord iagram or not.
Choose what annotations and track heights usong annotation.columns}
\usage{
circosPlot_cc_hca_tree(
  pdata = NULL,
  annotation.columns = c("hdbcl_gex", "taxonomy_published"),
  predefined.cc.column = "cc_gex",
  plot.name = "ChordPlot_GEX_cc_hdb_chen",
  plot.dir = "~/PROJECTS/RCC_ccpRCC_2019/reTaxonomy/2_Evaluate_cc_hdb/CircosPlots/",
 
    cc.dir = "~/PROJECTS/RCC_ccpRCC_2019/ConsensusClust/ConsensusCluster_rcc_t_880set_3177genes_reps.1000_maxK.25_innerLinkage.ward.D2_sd.1/",
  cc.k = 25,
  chord.column = NULL,
  chord.levels = c("outlier", "metanephric", "mtscRCC", "dCIMP", "mesRCC", "ccpRCC",
    "chONC", "pONC"),
  scale_factors = c(1, 3),
  color.key = NULL,
  dend.rev.clusters = NULL,
  dend.color.key = NULL,
  annotation.track.height = NULL
)
}
\arguments{
\item{pdata}{what pdata to use for plot}

\item{annotation.columns}{what pdata columns to plot as annotation segments}

\item{predefined.cc.column}{if the loaded cc already has been put in a specific pdata colum, e.g. cc_gex}

\item{plot.name}{name of file to plot}

\item{plot.dir}{directory to plot in}

\item{cc.dir}{directory in which consensus clust results are - parsed using dlfoo2::ConsensusCluster_process}

\item{cc.k}{number of consensusclusters (cannot be higher than performed in cc.dir)}

\item{chord.column}{name of pdata column if to plot add a second grup, pie section, to plot a chord plot.}

\item{chord.levels}{factor levels for the re-taxonomy. specifies order of groups in plot}

\item{scale_factors}{if to scale size of the chord part}

\item{color.key}{what color key for subgroups.}

\item{dend.rev.clusters}{if to flip clusters.}

\item{dend.color.key}{if a specific color key is to be used for dendogram}

\item{annotation.track.height}{what plot heights for the annotation tracks. integer vector of same length as annotation.columns}

\item{do.gradient}{set to FALSE if to manually override numeric check for gradient or manual fil}
}
\value{
plot pdf to file
}
\description{
specialized function alluvial chordPlot re-taxonomy w consensus HCA
}
\seealso{
Other chordplot: 
\code{\link{chordPlot_ccHCA_to_reTax}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other circlize: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{links2sectors}()},
\code{\link{rectMatrix}()}
}
\concept{PlotWrap}
\concept{chordplot}
\concept{circlize}
