% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_data_documentation.R
\docType{data}
\name{color_annotations}
\alias{color_annotations}
\title{Annotation color key for sample annotations}
\format{
Named character vector with colors
}
\usage{
color_annotations
}
\description{
Annotation color key for sample annotations
}
\keyword{datasets}
