% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_data_documentation.R
\docType{data}
\name{color_celltypes}
\alias{color_celltypes}
\title{Annotation color key for cell types}
\format{
Named character vector with colors
}
\usage{
color_celltypes
}
\description{
Annotation color key for cell types
}
\keyword{datasets}
