% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_data_documentation.R
\docType{data}
\name{color_subtypes}
\alias{color_subtypes}
\title{Color key for TCGA subtpes}
\format{
Named character vector with colors
}
\usage{
color_subtypes
}
\description{
Color key for TCGA subtpes
}
\keyword{datasets}
