% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{colorKeyFix}
\alias{colorKeyFix}
\title{Match up a color key to given a series of given unique annotation
Add colors if annots are missing}
\usage{
colorKeyFix(color.key = NULL, annots = NULL, na.color = "light gray")
}
\value{
color_key vector where all annotatoions have a color
}
\description{
Match up a color key to given a series of given unique annotation
Add colors if annots are missing
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other color: 
\code{\link{shapeKeyFix}()}
}
\concept{color}
\concept{plot wraps}
