% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{ConsensusCluster_process}
\alias{ConsensusCluster_process}
\title{wrapper that process resuts from ConsensusClusterPlus analyses}
\usage{
ConsensusCluster_process(cc.dir = NULL, k = 2)
}
\arguments{
\item{cc.dir, }{directory created by ConsensusClusterPlus analysis containing results. File must be called "cc.results.Rdata".}

\item{k, }{define the number of consensus cluster groups}
}
\value{
a list with 5 slots:
\itemize{
  \item cc: The absolute expression values over all genes present for the signature is averaged. Negatively correlated genes are weighted by (-1)
  \item i: Genes in the signature are treated as an expression vector based on their regulation (a series of +1 and -1, respectively). For each sample, the correlation between its expression values and the signature vector will represent the pathway score.
  \item consensus.tree: The consensus clutering tree
  \item mat: the consensus clust matrix representing co-clutering
  \item sample.names: sample names
}
}
\description{
Function used to process output data from \code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}} function.
}
\seealso{
Other hca: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{hca}()}

Other analysis: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()}
}
\concept{analysis}
\concept{hca}
