% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{ConsensusClusterPlus.dl}
\alias{ConsensusClusterPlus.dl}
\title{The consensusCluster-dl function}
\usage{
ConsensusClusterPlus.dl(
  d = NULL,
  run.name = NULL,
  results.dir = NULL,
  maxK = 3,
  reps = 10,
  innerLinkage = "ward",
  finalLinkage = "ward",
  distance = "pearson",
  var.method = "sd",
  var.cut = 1,
  pItem = 0.8,
  pFeature = 1,
  clusterAlg = "hc",
  ml = NULL,
  tmyPal = NULL,
  seed = NULL,
  plot = "png",
  writeTable = FALSE,
  weightsItem = NULL,
  weightsFeature = NULL,
  verbose = F,
  corUse = "everything"
)
}
\arguments{
\item{d}{expressionSet}

\item{run.name}{index vector that specify i) the number (k) of subclusters (length of vector), and ii) what subcluster that should come first, second ... etc}

\item{results.dir}{name of directory where analysis is to be saved}
}
\value{
consensus cluster results
}
\description{
Wrapper for ConsensusClusterPlus analysis \code{\link[ConsensusClusterPlus]{ccRun}}
}
\seealso{
Other hca: 
\code{\link{ConsensusCluster_process}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{hca}()}

Other analysis: 
\code{\link{ConsensusCluster_process}()},
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()}
}
\concept{analysis}
\concept{hca}
