% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Misc.R
\name{copyToMemory}
\alias{copyToMemory}
\title{copy a vetor to memory}
\usage{
copyToMemory(x, sep = "\\t")
}
\arguments{
\item{x}{vecor to be copied}

\item{sep}{what delim}
}
\value{
copied to memory to be pasted elsewhere
}
\description{
copy a vetor to memory
}
