% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Misc.R
\name{cormat2indiv}
\alias{cormat2indiv}
\title{Correlation matrix to table}
\usage{
cormat2indiv(cor.mat)
}
\arguments{
\item{cor.mat, }{a correlation matrix}
}
\value{
a data frame with all combinations of correlations
}
\description{
Uses a correlation martrix to create a table with correlations for all possible combinations
}
\seealso{
Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other transformation: 
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}

Other misc: 
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{misc}
\concept{old functions}
\concept{transformation}
