% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{corrCcle}
\alias{corrCcle}
\title{correlations in CCLE data set}
\usage{
corrCcle(
  ensg,
  geneSubset = NULL,
  tissue.site = c("prostate", "stomach", "urinary_tract", "central_nervous_system",
    "ovary", "haematopoietic_and_lymphoid_tissue", "kidney", "thyroid", "skin",
    "soft_tissue", "salivary_gland", "lung", "bone", "pleura", "endometrium", "pancreas",
    "breast", "upper_aerodigestive_tract", "large_intestine", "autonomic_ganglia",
    "oesophagus", "fibroblast", "liver", "biliary_tract", "small_intestine", "cervix",
    "haematopoietic_and_lymphoid", "buccal", "placenta", "eye", "adrenal_cortex")
)
}
\arguments{
\item{ensg}{ENSEMBL gene id}

\item{geneSubset}{A vector of ensg gene ids to limit the analysis}

\item{tissue}{tissue types to include in analysis. Defualts to "prostate","stomach","urinary_tract","central_nervous_system","ovary","haematopoietic_and_lymphoid_tissue","kidney","thyroid","skin","soft_tissue","salivary_gland","lung","bone","pleura","endometrium","pancreas","breast","upper_aerodigestive_tract","large_intestine","autonomic_ganglia","oesophagus","fibroblast","liver","biliary_tract","small_intestine","cervix","haematopoietic_and_lymphoid","buccal","placenta","eye","adrenal_cortex"}
}
\value{
correlation data frame
}
\description{
correlations in CCLE data set
}
\seealso{
Other analysis: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{corrGtex}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()}

Other gene expression: 
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other eset: 
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other ccle: 
\code{\link{geneplotCcle1156}()}
}
\concept{analysis}
\concept{ccle}
\concept{eset}
\concept{gene expression}
