% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{corrGtex}
\alias{corrGtex}
\title{correlation gene epxression in GTEX data set}
\usage{
corrGtex(
  ensg,
  tissue = c("AdiposeTissue", "AdrenalGland", "BloodVessel", "Bladder", "Brain",
    "Breast", "Blood", "Skin", "CervixUteri", "Colon", "Esophagus", "FallopianTube",
    "Heart", "Kidney", "Liver", "Lung", "SalivaryGland", "Muscle", "Nerve", "Ovary",
    "Pancreas", "Pituitary", "Prostate", "SmallIntestine", "Spleen", "Stomach", "Testis",
    "Thyroid", "Uterus", "Vagina")
)
}
\arguments{
\item{ensg}{ENSEMBL id}

\item{tissue}{what tossie types to include in analysis. Defautls to "AdiposeTissue","AdrenalGland","BloodVessel","Bladder","Brain","Breast","Blood","Skin","CervixUteri","Colon","Esophagus","FallopianTube","Heart","Kidney","Liver","Lung","SalivaryGland","Muscle","Nerve","Ovary","Pancreas","Pituitary","Prostate","SmallIntestine","Spleen","Stomach","Testis","Thyroid","Uterus","Vagina"}
}
\value{
table with correlations
}
\description{
correlation gene epxression in GTEX data set
}
\seealso{
Other analysis: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{corrCcle}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other eset: 
\code{\link{corrCcle}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gtex: 
\code{\link{geneplotGtex}()}
}
\concept{analysis}
\concept{eset}
\concept{gene expression}
\concept{gtex}
