% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{dend_classify}
\alias{dend_classify}
\title{dendextend package utils}
\usage{
dend_classify(dend, k)
}
\arguments{
\item{dend, }{an object of class \code{\link[dendextend]{dendrogram}}}

\item{k, }{the number of subclusters to be defined}
}
\value{
an integer vector with the same length as terminal nodes and that defines subcluster groups
}
\description{
Classify a dendogram into n=k subclusters using \code{\link[dendextend]{cutree}}. Can be used for input to reorder branches in a dendogram
}
\seealso{
Other hca: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{hca}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{hca}
\concept{transformation}
