% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{dend_reorder}
\alias{dend_reorder}
\title{dendextend package utils}
\usage{
dend_reorder(dend, cluster.order = c(2, 1))
}
\arguments{
\item{dend, }{an object of class \code{\link[stats]{dendrogram}}}

\item{cluster.order, }{index vector that specify i) the number (k) of subclusters (length of vector), and ii) what subcluster that should come first, second ... etc}
}
\value{
an object of class \code{\link[stats]{dendrogram}}
}
\description{
Reorder clusters within a dendrogram. Uses the \code{\link[dendextend]{rotate}} and \code{\link[dlfoo]{dend_classify}} functions.
}
\seealso{
Other hca: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{dend_classify}()},
\code{\link{dend_rev_clusters}()},
\code{\link{hca}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{hca}
\concept{transformation}
