% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{enrichmentAnalysis_ClusterProfiler}
\alias{enrichmentAnalysis_ClusterProfiler}
\title{enrichmentAnalysis_ClusterProfiler v2}
\usage{
enrichmentAnalysis_ClusterProfiler(
  gene.list = NULL,
  run.folder = NULL,
  run.name = NULL,
  entrez.or.symbol = "symbol",
  my_ontologies = c("h.all", "c2.cp.biocarta", "c2.cp.kegg", "c2.cp.reactome",
    "c3.tft", "c5.bp", "c5.cc", "c5.mf"),
  min.genes.in.term = 3
)
}
\arguments{
\item{gene.list, }{directory created by ConsensusClusterPlus analysis containing results. File must be called "cc.results.Rdata".}

\item{run.folder, }{name of higher level directory}

\item{run.name, }{name of run (directory created)}

\item{entrez.or.symbol}{if to run on suymbol or entrezgene}

\item{min.genes.in.term, }{minimum included genes in sigificant term}

\item{my.ontologies}{what ontologies to run}
}
\value{
write results to utput directory: EnricherResults.txt: tab delimited text file with all significant results from all terms and all gene lists. EnricherResults.rds: same as .txt but saved as .rds. EnricherResults_comparecluster.rds. ResultsSummary.txt: Results Summary on how many significant terms per gene list were obtained.
}
\description{
wrapper that perform Enrichment analysis using ClusterProfiler and DOSE packages

v2 that includes v7 of Msig 20190923
}
\seealso{
Other enrichment: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{enrichmentPlot_bubble}()},
\code{\link{gsea_RankList}()}

Other gene ontology: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{gsea_RankList}()}
}
\concept{enrichment}
\concept{gene ontology}
\concept{go}
