% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{enrichmentPlot_bubble}
\alias{enrichmentPlot_bubble}
\title{Generic plot for enrichment analyses}
\usage{
enrichmentPlot_bubble(
  ea.tab = NULL,
  score.title = "score",
  size.title = "size",
  order.by.clustering.x = T,
  order.by.clustering.y = T,
  border.col = NA,
  score.pal = c(rev(RColorBrewer::brewer.pal(9, "YlOrRd")[1:9]), rep("#FFFFCC", 2),
    rep("white", 5), rep("#FFFFCC", 2), RColorBrewer::brewer.pal(9, "YlGn")[1:9]),
  reverse.x = F,
  reverse.y = F
)
}
\arguments{
\item{ea.tab, }{data frame with enrichment analysis to plot.}

\item{score.title}{name of score value, e.g. p-value}

\item{size.title}{name of size value, e.g. p-value}

\item{order.by.clustering.x, }{let order of terms/groups be decided by clustering. Otherwise defined by levels}

\item{order.by.clustering.y, }{let order of terms/groups be decided by clustering. Otherwise defined by levels}

\item{border.col}{set to NA. If column named border is icluded then this column will be used}

\item{score.pal, }{color palette for p-val. if negative sizes (n).}

\item{reverse.x}{if to reverse x-axis}

\item{reverse.y}{if to reverse y-axis}
}
\value{
ggplot object
}
\description{
input is dataframe with, group, term, n, and p columns

https://bioconductor.org/packages/release/bioc/manuals/fgsea/man/fgsea.pdf

plot order is determined by clustering - if not then set term column to factor and let levels decide order

v2.0 20191016. fix pbug with color scales. Also move to plotting NES instead of pvalue ... and NES for ALL not only the significant.
}
\seealso{
Other enrichment: 
\code{\link{enrichmentAnalysis_ClusterProfiler}()},
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{gsea_RankList}()}

Other plotwrap: 
\code{\link{coOccurence_bubble}()}

Other msig: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{gsea_RankList}()}

Other gsea: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{geneRankPlot}()},
\code{\link{gsea_RankList}()}

Other fgsea: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{fgseaCoocPlotrWrap}()}
}
\concept{enrichment}
\concept{fgsea}
\concept{gsea}
\concept{msig}
\concept{plotwrap}
