% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_FeatureAnnotation.R
\name{entrez2chr}
\alias{entrez2chr}
\title{Get homo sapiens chromosome locations from a vector of entrez gene ids using org.Hs.eg.db pack}
\usage{
entrez2chr(entrez_id)
}
\arguments{
\item{entrez_id, }{a vector of entrez ids. May include NAs and duplicates.}
}
\value{
a feature data table with valid entrez ids and their respective chr
}
\description{
Get homo sapiens chromosome locations from a vector of entrez gene ids using org.Hs.eg.db pack
}
\seealso{
Other feature annotation: 
\code{\link{entrez2symbol}()},
\code{\link{featureGetBM}()},
\code{\link{features2genomicRanges}()},
\code{\link{homoloGene}()},
\code{\link{illuminav4_2fdata}()}

Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}
}
\concept{feature annotation}
\concept{old functions}
