% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_FeatureAnnotation.R
\name{entrez2symbol}
\alias{entrez2symbol}
\title{Get homo sapiens gene symbols from a vector of entrez gene ids using appropriate org.eg.db package.}
\usage{
entrez2symbol(entrez.id, keep.all.symbols = F, species = "Hs")
}
\arguments{
\item{keep.all.symbols, }{if to include entrez for which no sympol was found in output table.}

\item{species, }{define species (e.g. "Hs", "Mm", "Rn", "Dm"...). Defines what org.eg.db package that will be loaded.}

\item{entrez_id, }{a vector of entrez ids. May include NAs and duplicates.}
}
\value{
a feature data table with valid entrez ids (ENTREZID) and their respective gene symbols (SYMBOL)
}
\description{
Get homo sapiens gene symbols from a vector of entrez gene ids using appropriate org.eg.db package.
}
\seealso{
Other feature annotation: 
\code{\link{entrez2chr}()},
\code{\link{featureGetBM}()},
\code{\link{features2genomicRanges}()},
\code{\link{homoloGene}()},
\code{\link{illuminav4_2fdata}()}

Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}
}
\concept{feature annotation}
\concept{old functions}
