% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{esetCenterGenes}
\alias{esetCenterGenes}
\title{Center features/rows in an eset
Each row/feature in an eset will be re-rentered based on mean median or midpoint values. Can be performed on only a subset of samples.}
\usage{
esetCenterGenes(eset, measure = c("mean", "median", "max.min"), subset = NULL)
}
\arguments{
\item{eset, }{an object of class ExpressionSet}

\item{measure, }{can be 'mean','median', or 'max.min' (i.e., the midpoint between min and max values)}

\item{subset, }{specify sample names if measure should be calculated from only a subset of samples. Only these samples will get mean/median of 0.}
}
\value{
an ExpressionSet where each row now is centered to the mean, median or the max.min
}
\description{
Center features/rows in an eset
Each row/feature in an eset will be re-rentered based on mean median or midpoint values. Can be performed on only a subset of samples.
}
\seealso{
Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{eset}
\concept{gene expression}
\concept{transformation}
