% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{esetCreate}
\alias{esetCreate}
\title{Wrapper to Create a Biobase ExpressionSet
Create a Biobase \code{\link[Biobase]{ExpressionSet}} from a data matrix, fdata and pddta objects}
\usage{
esetCreate(
  data,
  pdata,
  fdata,
  sample.names.col = 1,
  feature.names.col = 1,
  platform = c(),
  data_processing = NULL
)
}
\arguments{
\item{data, }{a matrix containg expression values}

\item{pdata, }{data frame with phenotype data (pdata), i.e. sample information.}

\item{fdata, }{data frame with feature data (fdata), i.e. reporter/probe information}

\item{sample.names.col, }{index (number) of what column from pdata to use as sampleNames}

\item{feature.names.col, }{index (number) of what column from fdata to use as featureNames}

\item{platform, }{what platform is used - the annotation slot in the ExpressionSet object.}

\item{data_processing, }{additional information on data processing.}
}
\value{
an \code{\link[Biobase]{ExpressionSet}} object
}
\description{
Wrapper to Create a Biobase ExpressionSet
Create a Biobase \code{\link[Biobase]{ExpressionSet}} from a data matrix, fdata and pddta objects
}
\seealso{
Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{eset}
\concept{gene expression}
