% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{esetGexSignatures}
\alias{esetGexSignatures}
\title{function Plot GEX eset heatmaps to file - for signatures}
\usage{
esetGexSignatures(
  x.eset,
  signature.object,
  signature.annotation.column = "cell.type.tax",
  signature.id.type = "hgnc_symbol",
  eset.id.type = "SYMBOL",
  clusterRows = TRUE,
  sd.cut = 0.25,
  runName,
  runFolder,
  amapDist = c("pearson"),
  palNo = 3,
  palSat = c(-3, 3)
)
}
\arguments{
\item{x.eset}{filtered gene expression eSet with expression values to be plotted. Input shpuld be Sorted and Filtered!}

\item{signature.object}{object aproduct from featureGetBM analysis. A 'signature' slot should have been added.}

\item{signature.annotation.column}{defaults to 'cell.type.tax'. The annotation to break down the signature. Can be NULL if no subdivistion of the signature,}

\item{signature.id.type}{id type to match signature - gex eset. degfaults to 'hgnc_symbol'}

\item{eset.id.type}{SYMBOL default}

\item{clusterRows}{if to cluster signature rows (and each of the sub-signatuers)}

\item{runName}{name of plot/analyses - subfolder}

\item{runFolder}{directory where to plot/save objects}

\item{amapDist}{Supplied to amap::Dist function. Can be ONE of following distance methods. "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.}

\item{palNo}{What palette (dl_heat)}

\item{palSat}{saturation levels for palette. detfaults to '-3,3'}
}
\value{
heatmaps to file
}
\description{
Version 1.0 20190129
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other heatmap: 
\code{\link{esetHeatmapRCC}()},
\code{\link{heatmapAnnotations}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotHeatmap}()}

Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{eset}
\concept{gene expression}
\concept{heatmap}
\concept{plot wraps}
\concept{signature}
