% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{esetGroupMeans}
\alias{esetGroupMeans}
\title{Calculate sample group means for each row within an eset
Returns a matrix with mean values for each sample group and for each row (gene). Colums will be named after the annotation provide.}
\usage{
esetGroupMeans(x, annot, log2_data = F)
}
\arguments{
\item{x, }{an object of class ExpressionSet}

\item{annot, }{a vector that defines sample groups}

\item{log2.data, }{if to log2 the data}
}
\value{
a matrix with mean values for each sample group and for each row (gene). Colums will be named after the annotation provide and rows after featureNames of the eset
}
\description{
Calculate sample group means for each row within an eset
Returns a matrix with mean values for each sample group and for each row (gene). Colums will be named after the annotation provide.
}
\seealso{
Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{eset}
\concept{gene expression}
\concept{transformation}
