% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{esetHeatmapRCC}
\alias{esetHeatmapRCC}
\alias{gexHeatmapRCC}
\title{function Plot GEX eset heatmaps to file using standard Taxonomy.}
\usage{
esetHeatmapRCC(
  x.eset,
  runName,
  runFolder,
  clusterFeatures_k = 25,
  clusterSamples_k = 25,
  amapDist = c("pearson", "euclidean", "manhattan", "correlation"),
  palNo = 3,
  palSat = c(-3, 3)
)
}
\arguments{
\item{x.eset}{gene expression eSet with expression values to be plotted. Input shpuld be Sorted and Filtered!}

\item{runName}{name of plot/analyses - subfolder}

\item{runFolder}{directory where to plot/save objects}

\item{clusterFeatures_k}{How many subclusters should the gene HCA tree be split into}

\item{clusterSamples_k}{How many subclusters should the sample HCA tree be split into}

\item{amapDist}{Supplied to amap::Dist function. Can be one or more of distance methods. "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.}

\item{palNo}{What palette (dl_heat)}

\item{palSat}{saturation levels for palette. defaults to -3,3}
}
\value{
heatmaps to file
}
\description{
Performs HCA and heatmap plot using different distances

Version 1.0 20190129
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other heatmap: 
\code{\link{esetGexSignatures}()},
\code{\link{heatmapAnnotations}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotHeatmap}()}

Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{eset}
\concept{gene expression}
\concept{heatmap}
\concept{plot wraps}
