% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{esetMergeRows}
\alias{esetMergeRows}
\title{Merge (collapse) rows of an eset on replicate probes
Returns an eset where replicate probes are merged based on mean, median or highest sd probe}
\usage{
esetMergeRows(eset = NULL, fdata.column = NULL, method = "mean")
}
\arguments{
\item{eset, }{an object of class ExpressionSet}

\item{fdata.column, }{name of fData column in eset that specify what reporters to merge on}

\item{method, }{"mean" or "highest.sd"}
}
\value{
ExpressionSet for which replicate reporters are merged. fData columns are merged using pipe (|) when needed
}
\description{
Merge (collapse) rows of an eset on replicate probes
Returns an eset where replicate probes are merged based on mean, median or highest sd probe
}
\seealso{
Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}

Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}
}
\concept{eset}
\concept{gene expression}
\concept{old functions}
\concept{transformation}
