% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{expandRanges}
\alias{expandRanges}
\title{expand ranges up/downstream of a GRanges object
From https://support.bioconductor.org/p/78652/}
\usage{
expandRanges(x, upstream = 0, downstream = 0)
}
\arguments{
\item{x}{GRanges object}

\item{upstream}{no bases upstream to expand}

\item{downstream}{no bases downstream to expand}
}
\value{
GRanges object with expanded ranges from start stop as defined
}
\description{
expand ranges up/downstream of a GRanges object
From https://support.bioconductor.org/p/78652/
}
\seealso{
Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{genomics}
\concept{granges}
\concept{transformation}
