% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_FeatureAnnotation.R
\name{featureGetBM}
\alias{featureGetBM}
\title{Uses Biomart to retrieve gene info.}
\usage{
featureGetBM(
  query.id = "all_genes",
  query.type = c("ensembl_gene_id", "gene_symbol", "entrezgene_id"),
  query.species = c("hs", "mm", "rn"),
  hs.genome = c("hg38", "hg19"),
  mm.genome = c("mm10"),
  rn.genome = c("rn6"),
  keep.chromosomes = c(1:22, "X", "Y", "MT"),
  transcript.biotypes = c("protein_coding"),
  collapse.to.unique = TRUE,
  get_orthologs = FALSE
)
}
\arguments{
\item{query.id}{Input one or multiple gene identifiers as character vector (default is hgnc_gene symbol)}

\item{query.type}{What type of identifer: c("hgnc_symbol","entrezgene_id","ensembl_gene_id")}

\item{query.species}{What species (hs, mm, or rn)}

\item{hs.genome}{if to retrieve ENSG gene info from hg19 database instead of hg38. defauts to hg38}

\item{mm.genome}{only mm10 supported so far}

\item{rn.genome}{only rn6}

\item{keep.chromosomes}{Gebns on what chrosomosomes to keep. Defaults to c(1:22, "X","Y","MT"). set to NULL if to keep all genomic chromsomes, contigs and scaffolds}

\item{transcript.biotypes}{what transcript biotypes to keep. Defaults to 'protein_coding' only. To hkepp all, set to NULL.}

\item{collapse.to.unique}{if some input id's will generate nultiple output entries (e.g. one symbol may be associated with multiple ENSGs, e.e.  1:n). Should these be collapsed by piping or not?}
}
\value{
object list with slots
}
\description{
Use one gene/feature id to create object with multiple feature annotations, e.g. symbol, entrez, ENSG, and for multiple species

The featureIdObjectSignature is can be used in downstream analyses

For homologs - enseble is always used as primary id type when collapsing duplicate rows

NOTE!: update 20190916 . entrezgene changed to entrezgene_id
}
\seealso{
Other feature annotation: 
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{features2genomicRanges}()},
\code{\link{homoloGene}()},
\code{\link{illuminav4_2fdata}()}
}
\concept{biomart}
\concept{feature annotation}
