% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{fromChr}
\alias{fromChr}
\title{Switch from enembl to UCSC stype chromosome nomenclature}
\usage{
fromChr(chr)
}
\arguments{
\item{chr}{object of class GenoSet}
}
\value{
numeric vector
}
\description{
Switch from enembl to UCSC stype chromosome nomenclature
}
\seealso{
Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{genomics}
\concept{misc}
\concept{old functions}
