% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{geneplotTcgaRccTax}
\alias{geneplotTcgaRccTax}
\title{boxplot single gene epxression in TCGA urinary tract (UTC) data}
\usage{
geneplotTcgaRccTax(ensg = NULL, ...)
}
\arguments{
\item{ensg}{A ensemble gene identifier (homo sapiens)}
}
\value{
a plot
}
\description{
boxplot single gene epxression in TCGA urinary tract (UTC) data
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()},
\code{\link{tcga_shortname}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{plotHeatmap}()},
\code{\link{tcgaSurvPlot2}()}

Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{eset}
\concept{gene expression}
\concept{plot wraps}
\concept{tcga}
