% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{geneRankPlot}
\alias{geneRankPlot}
\title{Enrichemnt plot in gsea-line format. One line/vertical bar for each gene
Genes are provided as a list of named vectors vit values (different groups)
gene values are given a rank and plotted (x-axis). on y-axis the different groups are ordered (acornding to mean rank)
rank lines are colored and according to color.key provided}
\usage{
geneRankPlot(
  x,
  gene.id.column = "SYMBOL",
  group.column = "contrast",
  value.column = "t.value",
  rank.descending = T,
  n_genes = NULL,
  pathway.genes = NULL,
  pathway.name = NULL,
  reverse.y = T,
  color.gradient.on.rank = T,
  ggplotply.prepare = F,
  color.key = dlfoo2::color_annotations,
  ...
)
}
\arguments{
\item{x}{Data frame with, at least:  one 'group', one 'value', and one 'gene_id' column}

\item{gene.id.column}{specify name of gene.id column}

\item{group.column}{specify name of value column - used to create RANKS}

\item{value.column}{specify name of group column}

\item{rank.descending}{TRUE if leftmost should be the highest x value}

\item{n_genes}{the max rank (x-value)}

\item{pathway.genes}{Vector of genes that belong to a specific pathway to plot}

\item{pathway.name}{I to add statistics}

\item{reverse.y}{if to reverse y axis}

\item{color.gradient.on.rank}{if to color each group (row) on a group color a gradient based on the rank value. Defaults to TRUE}

\item{ggplotply.prepare}{if output is for ggplotly, then set this to TRUE to get a ggploty gradient (ggplotly uses fill_ instead of color_)}

\item{color.key}{color key for group colors if gradient is FALSE}
}
\value{

}
\description{
Enrichemnt plot in gsea-line format. One line/vertical bar for each gene
Genes are provided as a list of named vectors vit values (different groups)
gene values are given a rank and plotted (x-axis). on y-axis the different groups are ordered (acornding to mean rank)
rank lines are colored and according to color.key provided
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other gsea: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{enrichmentPlot_bubble}()},
\code{\link{gsea_RankList}()}
}
\concept{gsea}
\concept{plot wraps}
