% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{genomeCircleSegments}
\alias{genomeCircleSegments}
\title{Wrapper for Plotting (Multiple) genome tracks in Circlize -- One sample multiple tracks.}
\usage{
genomeCircleSegments(
  x = my_cn,
  plot.mode = c("single.track.type"),
  plot.name = "",
  track.types = c("rect"),
  value.column = c("score"),
  genome.version = "hg38",
  track.height = NULL,
  palette = list(rev(c(rev(RColorBrewer::brewer.pal(9, "YlOrRd")[c(3:9, 9)]),
    rep("white", 1), RColorBrewer::brewer.pal(9, "Blues")[2:9]))),
  palette.sat = list(c(-0.75, 0.75)),
  my.chr = paste0("chr", c(1:22, "X")),
  sort.pos = c("chr3", "10141635"),
  my.lwd = NULL,
  label.bed = NULL,
  label.side = "inside",
  label.cex = 0.8,
  ideogram.plotType = c("ideogram", "axis", "labels"),
  ideogram.ideogram.height = convert_height(2, "mm")
)
}
\arguments{
\item{x}{GenomicRanges or bed style data.frame  (if one 1 sample) or list of genomic ranges objects OR a BED style data.frames (if multiple samples or if multiple tracks). Chromosomes should be in 'chrX' format (not 'X')}

\item{plot.mode}{If x is a list and if multiple tracks (not multiple samples for a single track) is to be plotted. If multiple tracks set to "multiple.track.types". Default is "single.track.type".}

\item{plot.name}{name of plot}

\item{track.types}{What type of plot connected to each of the list objects. "rect" or "line.bar" supported. Character vector of length(x) if multiple and different tracks.}

\item{value.column}{name of the column in bed/GRanges that will define segment colors ('score' is default)}

\item{genome.version}{Defaults to hg38. accepts multiple species}

\item{track.height}{Numeric vector. If different track.heights then this should be the length of x.}

\item{palette}{list of cahacter vectors. the palette(s) to use for the different tracks. If tracks are not numeic, then names of palette will be used to define colors.}

\item{palette.sat}{palette sat}

\item{my.chr}{vector with what chromosomes to plot}

\item{sort.pos}{what chromosome position to sort on (chr, start, end)}

\item{my.lwd}{line width for segments}

\item{label.bed}{bed style data.frame. if to label genes/regions in plot. Must include 'chr' (in chrX format), 'start' 'end' and 'label' column. May include 'color' column if user-defined colors.}

\item{ideogram.plotType}{for circos.initializeWithIdeogram function. One, tow or all of "ideogram", "axis", "labels"}

\item{ideogram.ideogram.height}{Height of the ideogram track}

\item{ideogram.track.height}{Height of the track which contains "axis" and "labels"}
}
\value{
genomeCircleSegments
}
\description{
wrapper for circlize function described at https://jokergoo.github.io/circlize_book/book/modes-of-input.html

version 2.0. accepts multiple track types

Plot genome data from BED/genomic ranges objects.

Input can be GenomicRanges object (if one 1 sample) or list of genomic ranges objects OR a BED style data.frame. Columns must include "ID" if multiple samples, "chr" if bed.
}
\seealso{
Other circlize: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{links2sectors}()},
\code{\link{rectMatrix}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other copy number: 
\code{\link{genomeCircleSegments_legacy}()}

Other genomic ranges: 
\code{\link{genomeCircleSegments_legacy}()}
}
\concept{PlotWrap}
\concept{circlize}
\concept{copy number}
\concept{genomic ranges}
