% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{genomeCircleSegments_legacy}
\alias{genomeCircleSegments_legacy}
\title{Wrapper for Plotting Copy number prfiles of Chrs in Circlize. each sample represented by a segment}
\usage{
genomeCircleSegments_legacy(
  x,
  plot.name = "temp",
  my.chr = paste0("chr", c(1:22, "X")),
  genome.version = "hg38",
  sort.pos = c("chr3", "10141635"),
  value.column = "score",
  palette = rev(c(rev(RColorBrewer::brewer.pal(9, "YlOrRd")[2:9]), rep("white", 3),
    RColorBrewer::brewer.pal(9, "Blues")[2:9])),
  palette.sat = c(-0.75, 0.75),
  my.lwd = NULL,
  label.bed = NULL,
  label.side = "inside",
  label.cex = 0.8,
  track.height = 0.3
)
}
\arguments{
\item{x}{GenomicRanges object (if one 1 sample) or list of genomic ranges objects OR a BED style data.frame}

\item{plot.name}{name of file to plot}

\item{my.chr}{vector with what chromosomes to plot}

\item{genome.version}{Defaults to hg38. accepts multiple species}

\item{value.column}{name of column in bed/granges to define copy numnber colors (score)}

\item{palette}{palette}

\item{palette.sat}{palette sat}

\item{my.lwd}{line width for segments}

\item{label.bed}{bed style data.frame. if to label genes/regions in plot. Must include 'label' column. May include 'color' column if user-defined colors.}

\item{plot.dir}{directory to plot in}
}
\value{
plot circle
}
\description{
wrapper for circlize function described at https://jokergoo.github.io/circlize_book/book/modes-of-input.html

Plot genome data from BED/genomic ranges objects.

Input can be GenomicRanges object (if one 1 sample) or list of genomic ranges objects OR a BED style data.frame. Columns must include "ID" if multiple samples, "chr" if bed.
}
\seealso{
Other circlize: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{genomeCircleSegments}()},
\code{\link{links2sectors}()},
\code{\link{rectMatrix}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other copy number: 
\code{\link{genomeCircleSegments}()}

Other genomic ranges: 
\code{\link{genomeCircleSegments}()}
}
\concept{PlotWrap}
\concept{circlize}
\concept{copy number}
\concept{genomic ranges}
