% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{genomicPosition}
\alias{genomicPosition}
\title{get full (accumulated) genomic position}
\usage{
genomicPosition(pos, chrom, force.full.genome.positions = F, add.space = 0)
}
\arguments{
\item{pos}{vector of genomic position(s)}

\item{chrom}{vector of chromosomes}

\item{force.full.genome.positions}{if to add virual (cumulative) genome positions}

\item{add.space}{if to add space in between chromosomes}
}
\value{
gemomic locations
}
\description{
get full (accumulated) genomic position
}
\seealso{
Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{homoloGene}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{genomics}
\concept{misc}
\concept{old functions}
