% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_ImportExport.R
\name{geoSeriesMatrixPdataPolish}
\alias{geoSeriesMatrixPdataPolish}
\title{small script for polisihing pdata seriesmatrix from geo}
\usage{
geoSeriesMatrixPdataPolish(pdata = NULL)
}
\arguments{
\item{pdata}{pdata from geo series matrix entry}
}
\description{
Wrapper for importing soft data from GEO using the GEOquery package

Methylation 450 data only

will download the full soft file and thereafter extract info
}
\seealso{
Other GEO: 
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseIDAT}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other series matrix: 
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genoset: 
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other import: 
\code{\link{get.HPA.images}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}
}
\concept{GEO}
\concept{genoset}
\concept{import}
\concept{series matrix}
