% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_ImportExport.R
\name{get.HPA.images}
\alias{get.HPA.images}
\title{get.HPA.images}
\usage{
get.HPA.images(
  symbol = NULL,
  run.name = NULL,
 
    hpa.filePaths.file = "~/RESOURCES/HPR/proteinatlas_v13_2011127.xml_all_tissues_normal_filePaths.Rdata",
  tissue.type = "kidney",
  hpa.images.dir = "/Volumes/MacPro2TB//RAW DATA/HPA/HPA_IMAGES/",
  out.dir = "/Volumes//MacPro2TB//RAW DATA/HPA/HPA_images_geneSignatures/",
  copy.to.out.dir = F,
  plot.to.file = T
)
}
\arguments{
\item{symbol}{gene symbol}

\item{run.name}{name of this run. needed to generate file names}

\item{hpa.filePaths.file}{file where HPA filepaths is stored. defaults to ~/RESOURCES/HPR/proteinatlas_v13_2011127.xml_all_tissues_normal_filePaths.Rdata}
}
\value{
save images to disk
}
\description{
Wrapper for importing Human Protein atlas images
}
\seealso{
Other import: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}
}
\concept{human protein atlas}
\concept{import}
