% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{ggPlotEmpty}
\alias{ggPlotEmpty}
\title{ggPlotEmpty}
\usage{
ggPlotEmpty(main = NULL)
}
\arguments{
\item{main}{plot main}
}
\value{
empty ggplot object
}
\description{
Version 1.0 20190110
}
\seealso{
Other plot functions: 
\code{\link{heatmap.3}()},
\code{\link{palette_gradientRamp}()}
}
\concept{plot functions}
