% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{gsea_RankList}
\alias{gsea_RankList}
\alias{gsea.rnk.object}
\title{create a rank-list as input for GSEA}
\usage{
gsea_RankList(
  gene.identifiers,
  rank.values,
  invert.rank = FALSE,
  collapse.method = "highest rank",
  convert.input.to.rank = FALSE,
  use.absolute.for.triming = TRUE
)
}
\arguments{
\item{gene.identifiers, }{list of named numeric vectors. One for each entity (sample/group etc). Each numeric vector is named with gene identifiers. no duplicated (duplicated names allowed)}

\item{rank.values, }{name of higher level directory}

\item{invert.rank, }{name of run (directory created)}

\item{collapse.method, }{what mSig or other signatues to load}

\item{convert.input.to.rank}{fgsea param min size}

\item{use.absolute.for.triming}{fgsea param max size}
}
\description{
create a rank-list as input for GSEA
}
\seealso{
Other enrichment: 
\code{\link{enrichmentAnalysis_ClusterProfiler}()},
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{enrichmentPlot_bubble}()}

Other gene ontology: 
\code{\link{enrichmentAnalysis_ClusterProfiler}()},
\code{\link{enrichmentAnalysis_fgsea}()}

Other msig: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{enrichmentPlot_bubble}()}

Other gsea: 
\code{\link{enrichmentAnalysis_fgsea}()},
\code{\link{enrichmentPlot_bubble}()},
\code{\link{geneRankPlot}()}
}
\concept{enrichment}
\concept{gene ontology}
\concept{gsea}
\concept{msig}
