% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{hca}
\alias{hca}
\title{Wrapper for HCA}
\usage{
hca(
  x,
  cor.method = "pearson",
  linkage = "ward.D",
  plot.dendogram = TRUE,
  sample.names = NULL,
  my.plot.args = list(cex = 0.5, main = NULL),
  new = FALSE,
  mar = c(4, 4, 2, 1),
  hang = -1
)
}
\arguments{
\item{x, }{a matrix to cluster samples or genes}

\item{cor.method, }{can be one of 'pearson', 'kendall', 'spearman', 'jacard', or 'euclidean'}

\item{linkage, }{default is 'ward.d'. as defined by the \code{\link[stats]{hclust}} function in stats package.}

\item{plot.dendogram, }{if to plot HCA tree. Default is TRUE.}

\item{sample.names, }{a vector that will replace sample names.}

\item{my.plot.args}{a list with graphical arguments.}
}
\value{
a list with two slots: $hr: reults object of class hclust and $d, the distance matrix
}
\description{
Perform Hierarchical Cluster Aanalysis using \code{\link[stats]{hclust}} function in stats package.
}
\seealso{
Other hca: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{hca}
\concept{transformation}
