% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{heatcol_creator}
\alias{heatcol_creator}
\title{create heat-color matrix from eset or matrix}
\usage{
heatcol_creator(
  x,
  palette.sat = c(-1, 1),
  pal = c("blue.white.red", "blue.white.yellow", "blue.black.red", "blue.black.yellow",
    "white.red", "white.blue", "white.brown", "white.redbrown", "white.green",
    "white.black"),
  na.col = "gray",
  low.mid.high = c(1, 50, 100)
)
}
\arguments{
\item{x, }{an expressionset or a matrix}

\item{palette.sat, }{vector with two values that define saturation levels c(min, max)}

\item{pal, }{define what palette to be used. Defined from object "palette.dl.heatcol"}

\item{na.col, }{color for na-values}

\item{low.mid.high, }{if to skew midpoint of color scale. vector with three values (percent), c(1,50,100).}
}
\value{
a list with two slots; col, a matrix with colors; legend, vector with colors mapped to their respecive values
}
\description{
\code{heatcol_creator}
Function to create a matrix with hexadecimal colors used for heatmaps
}
