% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{heatcol_creator2}
\alias{heatcol_creator2}
\title{create heat-color matrix from eset or matrix - alternate version}
\usage{
heatcol_creator2(
  x,
  palette.sat = c(-1, 1),
  my.pal = RColorBrewer::brewer.pal(9, "YlOrRd"),
  na.col = "gray",
  n.col = 50
)
}
\arguments{
\item{x, }{an expressionset or a matrix}

\item{palette.sat, }{vector with two values that define saturation levels c(min, max)}

\item{my.pal, }{a vector withcolors. The palette will be split on these in "equal" parts}

\item{na.col, }{color for na-values}

\item{n.col, }{the number of color steps}
}
\value{
a list with two slots; col, a vector/matrix with colors; legend, vector with colors mapped to their respecive values
}
\description{
\code{heatcol_creator2}
Function to create a matrix with hexadecimal colors used for heatmaps
}
