% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_FeatureAnnotation.R
\name{homoloGene}
\alias{homoloGene}
\title{use locally stored NCBI homologene data (homologene.data.build68.Rdata) to find inter-species homologs gene IDs}
\usage{
homoloGene(entrez.id, from.tax = "Mm", to.tax = "Hs", keep.all.hids = T)
}
\arguments{
\item{entrez.id, }{entrez_id}

\item{from.tax, }{from what species (Mm, Hs, Rn, Dm)}

\item{to.tax, }{from what species (Mm, Hs, Rn, Dm)}

\item{keep.all.hids, }{logical if to keep}
}
\value{
a data frame with from and to entrezid and homolgy ID - HID
}
\description{
use locally stored NCBI homologene data (homologene.data.build68.Rdata) to find inter-species homologs gene IDs
}
\seealso{
Other feature annotation: 
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{featureGetBM}()},
\code{\link{features2genomicRanges}()},
\code{\link{illuminav4_2fdata}()}

Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{loadRefgene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}
}
\concept{feature annotation}
\concept{old functions}
