% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_FeatureAnnotation.R
\name{illuminav4_2fdata}
\alias{illuminav4_2fdata}
\title{update Illumina v4 probe ids using the IlluminaHumanv4.db package}
\usage{
illuminav4_2fdata(probe_ids)
}
\arguments{
\item{probe_ids, }{vector containing valid Illumina probe IDs}
}
\value{
a data frame with feature data
}
\description{
update Illumina v4 probe ids using the IlluminaHumanv4.db package
}
\seealso{
Other feature annotation: 
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{featureGetBM}()},
\code{\link{features2genomicRanges}()},
\code{\link{homoloGene}()}
}
\concept{feature annotation}
