% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{label_wrap_mod}
\alias{label_wrap_mod}
\title{Wrap Text in ggplots.
Function to wrap text with newline. Used in ggplots (e.g. enrichmentPlot_bubble)}
\usage{
label_wrap_mod(value, width = 15)
}
\arguments{
\item{value, }{a matrix}

\item{width, }{how many characters are accepted per line}
}
\value{
, text wrapped with newline
}
\description{
Wrap Text in ggplots.
Function to wrap text with newline. Used in ggplots (e.g. enrichmentPlot_bubble)
}
\seealso{
Other ggplot: 
\code{\link{boxplot_SHINY_annotation}()}
}
\concept{ggplot}
\concept{plotmisc}
