% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{liftoverGRanges}
\alias{liftoverGRanges}
\title{Wrapper Liftover using AnnotaitonHub package
Using AnnotaionHub  \code{\link[AnnotationHub]{AnnotationHub}} to perform liftover between genome bulds.}
\usage{
liftoverGRanges(
  x = NULL,
  id_type = c("ensembl_gene_id", "hgnc_symbol", "entrezgene"),
  to = c("hg19"),
  keep.chromosomes = c(1:22, "X")
)
}
\arguments{
\item{x}{A \code{\link[GenomicRangges]{GRanges}} object. x must have genome specified. all sequnces must be from one unique genome.}

\item{to}{what genome build to transfer to. Default is hg19}

\item{keep.chromosomes}{what chrosomosomes to keep. set to NULL if to keep all chromosomes}

\item{collapse.to.unique}{If output should contain uniqe ids (per row) or if multiple rows are allowed for out. The non-collapsed table will be stored in metadata slot}
}
\value{
GRanges object with genomic start/stop for supplied identifiers (or all if 'all' option used)
}
\description{
Wrapper Liftover using AnnotaitonHub package
Using AnnotaionHub  \code{\link[AnnotationHub]{AnnotationHub}} to perform liftover between genome bulds.
}
\seealso{
Other methylation: 
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{genomics}
\concept{genoset}
\concept{granges}
\concept{methylation}
\concept{transformation}
