% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{links2sectors}
\alias{links2sectors}
\title{Small function to extract x (left and right) values between two groups for plotting in circlize (cicos.link) functions.
if multiple links from (and to) segments - then create x values as cumsum so links do not start/end at same positions
Used as auxillary funtion to be able to 'zoom' which is not possible in chordDiagram()}
\usage{
links2sectors(links.df, scale.x1 = 1, scale.x2 = 1, x.offset = 0)
}
\arguments{
\item{links.df}{A data frame with connections between two groups.}

\item{scale.x1}{if x-axis (no connections should be scaled)}

\item{scale.x2}{if y-axis (no connections should be scaled)}

\item{x.offset}{if to offset x-values}
}
\value{
Data frame with connected x-values to use in links function (x1 and x2 slots)
}
\description{
Small function to extract x (left and right) values between two groups for plotting in circlize (cicos.link) functions.
if multiple links from (and to) segments - then create x values as cumsum so links do not start/end at same positions
Used as auxillary funtion to be able to 'zoom' which is not possible in chordDiagram()
}
\seealso{
Other plots: 
\code{\link{rectMatrix}()}

Other circlize: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{rectMatrix}()}
}
\concept{circlize}
\concept{circos.link}
\concept{plots}
