% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_GenomicWraps.R
\name{loadRefgene}
\alias{loadRefgene}
\title{Load local UCSC refGene object corresponding to the genome of input genomic Ranges object
Load local UCSC refGene object corresponding to the genome of input genomic Ranges object. Uses the genome slot}
\usage{
loadRefgene(gr.obejct)
}
\arguments{
\item{gr.obejct}{a genomic ranges object whose matching refgene to be loaded}
}
\value{
a granges object
}
\description{
Load local UCSC refGene object corresponding to the genome of input genomic Ranges object
Load local UCSC refGene object corresponding to the genome of input genomic Ranges object. Uses the genome slot
}
\seealso{
Other old functions: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{cormat2indiv}()},
\code{\link{entrez2chr}()},
\code{\link{entrez2symbol}()},
\code{\link{esetMergeRows}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{homoloGene}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}
}
\concept{genomics}
\concept{granges}
\concept{old functions}
