% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{me2gexHeatmapRCC}
\alias{me2gexHeatmapRCC}
\title{Plot Methylation Beta heatmap with matched GEX heatmao}
\usage{
me2gexHeatmapRCC(
  x.me,
  x.gex,
  runName,
  runFolder,
  clusterFeatures_k = 25,
  clusterSamples_k = 25,
  amapDist = c("pearson", "euclidean", "manhattan", "correlation"),
  palNo = 3,
  palSat = c(0, 1),
  palNoGex = 1,
  palSatGex = c(-3, 3)
)
}
\arguments{
\item{x.me}{methylation GenoSet with 'beta' slot}

\item{runName}{name of plot/analyses - subfolder}

\item{runFolder}{directory where to plot/save objects}

\item{clusterFeatures_k}{number of clusters for CpG feature tree}

\item{clusterSamples_k}{number of clusters for sample tree}

\item{amapDist}{Supplied to amap::Dist function. Can be one or more of distance methods. "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.}

\item{palNo}{What palette (dl_heat)}

\item{palSat}{saturation levels for beta levels. defaults to '0,1'}

\item{palNoGex}{numer (id) of palette to use. uses dlfoo2::palette_dl_heatmap2}

\item{x.eset}{expressionSet with gex data}

\item{paSatGex}{saturatuon levels for gene expression. defaults to '0,3'}
}
\value{
plots to file
}
\description{
Version 1.0 20190124

Function to perform a standard HCA and heatmap plot of meatched me/gex data using RCC (and bladder) taxonomy. Saves .png, .pdf and cluster trees (.rds) to file. Requires genoset with ECR tab as input

ECR tab describes Me-Gex pairs. And lists correlaions, Extension (and replaces meHeatmapRCC but adds also a GEX heatmap
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other heatmap: 
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{heatmapAnnotations}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotHeatmap}()}

Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other eset: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{eset}
\concept{genoset}
\concept{heatmap}
\concept{methylation}
\concept{plot wraps}
