% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Methylation.R
\name{meAnalysisCorrGexRandom}
\alias{meAnalysisCorrGexRandom}
\title{randomize correlation pairs CpG and GEX}
\usage{
meAnalysisCorrGexRandom(
  x.me,
  x.gex,
  runName,
  runFolder,
  gexSDcut = 0.25,
  corWindow = 1e+06,
  mySeed = 100,
  nRandomPairs = 5e+06
)
}
\arguments{
\item{x.me}{object of class GenoSet. Contains assayData slot 'beta' with beta values. Sample names (colnames) must match sampleNames for x.gex}

\item{x.gex}{object of class expressionSet. Contains gene expression data. sampleNames for x.gex must match sample names (colnames) for x.me}

\item{runName}{Fie/dir-Name to use forthis run, e.g. rcc_me450.}

\item{runFolder}{Higher level directory where processed genoset object should be saved to}

\item{gexSDcut}{Variance filter cutoff for GEX data. Defaults to 0.5.}

\item{corWindow}{window surrounding CpG to include genes for correlation analysis}

\item{mySeed}{Seed number for random sampling}

\item{nRandomPairs}{Number of random Me-GEX pairs to be sampled}
}
\value{
saving table of correlations to file
}
\description{
Version 1.0 20190110

Function:  Correlation Me beta vs GEX - RANDOM SAMPLING
}
\seealso{
Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other analysis: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()}

Other filter: 
\code{\link{esetVarianceFilter}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()}
}
\concept{analysis}
\concept{filter}
\concept{genomics}
\concept{genoset}
\concept{granges}
\concept{methylation}
