% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Methylation.R
\name{meAnalysisCorTableStats}
\alias{meAnalysisCorTableStats}
\title{meAnalysisCorTableStats}
\usage{
meAnalysisCorTableStats(
  meAnalysisCorrGexFolder,
  meAnalysisCorrGexFolderRandom,
  x.me,
  densAdjust = 1/5,
  myQuantiles = c(0.001, 0.999),
  returnPlots = FALSE
)
}
\arguments{
\item{meAnalysisCorrGexFolder}{Path to directory where Me-GEX Correlation Analysis was performed}

\item{meAnalysisCorrGexFolderRandom}{Folder path to random subsampling data.}

\item{x.me}{methylation genoset with 'beta' values}

\item{densAdjust}{Adjustment value for density plots}

\item{myQuantiles}{Quantiles for random correlation analyses to define correlation cutoffs.}

\item{returnPlots}{if to retrun list of ggplot obejects}
}
\value{
Returns tables and pdfs to file
}
\description{
Version 1.0 20190115

Uses methylation genoset from a Me-GEX Correlation Analysis of VMPs \code{\link[dlfoo2]{meAnalysisCorrGex}}

Used to select signficant Expression Correlated Refions (ECRs). Additional inpt must be a run with random generated me::gex correlation pairs used to define sgnificance. from \code{\link[dlfoo2]{meAnalysisCorrGexRandom}}
}
\seealso{
Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other granges: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{features2genomicRanges}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meMergeCpGs}()},
\code{\link{segmentPlotter}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other filter: 
\code{\link{esetVarianceFilter}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()}
}
\concept{filter}
\concept{genomics}
\concept{genoset}
\concept{granges}
\concept{methylation}
\concept{stats}
