% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_ImportExport.R
\name{meGetGEOsoft_me450}
\alias{meGetGEOsoft_me450}
\title{Download and parse GEO .soft methylation data into GenoSet}
\usage{
meGetGEOsoft_me450(
  geo.id = NULL,
  runName = NULL,
  runFolder = "/Volumes/MacPro2TB/RESOURCES/Methylation_Infinium/Raw_data/",
  projectLabel = NULL,
  NormalizeBeta = F,
  platform = "GPL13534"
)
}
\arguments{
\item{geo.id}{GEO GSE id}

\item{runName}{Fie dir-Name to use forthis run, e.g. rcc_me450.}

\item{runFolder}{Higher level directory to store processed genoset object}

\item{projectLabel}{project lebel to be added as metadata(genoSet)}

\item{NormalizeBeta}{if to normalize beta infimium probe sets ringner/staaf style}

\item{platform}{specifiec what platform to extract. defaults to 450k GPL13534}
}
\value{
save tables to disk
}
\description{
Wrapper for importing soft data from GEO using the GEOquery package

Methylation 450 data only

will download the full soft file and thereafter extract info
}
\seealso{
Other GEO: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{meParseIDAT}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other import: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{get.HPA.images}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}
}
\concept{GEO}
\concept{genoset}
\concept{import}
\concept{methylation}
