% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{meHeatmapRCC}
\alias{meHeatmapRCC}
\title{Plot Methylation Beta heatmap - standard}
\usage{
meHeatmapRCC(
  x.me,
  runName,
  runFolder,
  clusterFeatures_k = 25,
  clusterSamples_k = 25,
  amapDist = c("pearson", "euclidean", "manhattan", "correlation"),
  palNo = 3,
  palSat = c(0, 1)
)
}
\arguments{
\item{x.me}{GenoSet object with 'beta' slot}

\item{runName}{name of plot/analyses - subfolder}

\item{runFolder}{directory where to plot/save objects}

\item{clusterFeatures_k}{number of clusters for CpG feature tree}

\item{clusterSamples_k}{number of clusters for sample tree}

\item{amapDist}{Supplied to amap::Dist function. Can be one or more of distance methods. "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.}

\item{palNo}{What palette (dl_heat)}

\item{palSat}{saturation levels for palette, defaults to '(0,1)'}
}
\value{
plots to file
}
\description{
Version 1.0 20190124

Function to perform a standard HCA and heatmap plot using RCC (and bladder) taxonomy. Saves .png, .pdf and cluster trees (.rds) to file.
}
\seealso{
Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other heatmap: 
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{heatmapAnnotations}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()}

Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}
}
\concept{PlotWrap}
\concept{genoset}
\concept{heatmap}
\concept{methylation}
