% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Methylation.R
\name{meMergeGenoSetsBeta}
\alias{meMergeGenoSetsBeta}
\title{meMergeGenoSetsBeta}
\usage{
meMergeGenoSetsBeta(gs_list)
}
\arguments{
\item{gs_list}{list of objects of class genoset}
}
\value{
save rds to file
}
\description{
Version 1.0 20190209
}
\seealso{
Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}
}
\concept{genomics}
\concept{genoset}
\concept{merge}
\concept{methylation}
