% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Methylation.R
\name{meNormalizeBeta}
\alias{meNormalizeBeta}
\title{Normalize Illumina Methylation Ringner/Staaf}
\usage{
meNormalizeBeta(
  x,
  runName,
  runFolder,
  returnObject = F,
  projectLabel = NULL,
 
    probe.annotations.file = "~/RESOURCES/Methylation_Infinium/Annotations/Probe.manifest_masked_intersect_450.epic_hg38.rds"
)
}
\arguments{
\item{x}{object of class GenoSet}

\item{runName}{Fie dir-Name to use forthis run, e.g. rcc_me450.}

\item{runFolder}{directory where processed genoset object should be saved to}

\item{returnObject}{if to return object instead of write to disk}

\item{projectLabel}{character vecto. name of project}

\item{probe.annotations.file}{file name for probe annotations that inlcude 'INFINIUM_DESIGN_TYPE' column: type I and II ILMN probe sets}
}
\value{
genoset
}
\description{
function meNormalizeBeta

Version 1.0 20190108
}
\seealso{
Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other genomics: 
\code{\link{chip_genePeaks2occupancyTab}()},
\code{\link{chip_peak2gene_maxValue}()},
\code{\link{chip_peak2gene}()},
\code{\link{expandRanges}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{liftoverGRanges}()},
\code{\link{loadRefgene}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{segmentPlotter}()},
\code{\link{toChr}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{dend_rev_clusters}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()}
}
\concept{genomics}
\concept{genoset}
\concept{methylation}
\concept{transformation}
