% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_ImportExport.R
\name{meParseIDAT}
\alias{meParseIDAT}
\title{Download and parse GEO .soft methylation data into GenoSet}
\usage{
meParseIDAT(
  pdata = NULL,
  pdata.sampleName.column = "geo_accession",
  idatFolder = NULL,
  runName = NULL,
  runFolder = "/Volumes/MacPro2TB/RESOURCES/Methylation_Infinium/Raw_data/",
  projectLabel = NULL,
  NormalizeBeta = F
)
}
\arguments{
\item{pdata}{sample sheet with idat file names. used as pdata.}

\item{pdata.sampleName.column}{column to use for sample names. must be unique}

\item{idatFolder}{top level folder for IDATs}

\item{runName}{Fie dir-Name to use forthis run, e.g. rcc_me450.}

\item{runFolder}{Higher level directory to store processed genoset object}

\item{projectLabel}{project lebel to be added as metadata(genoSet)}
}
\value{
save tables to disk
}
\description{
Wrapper for importing soft data from GEO using the GEOquery package

Methylation 450 data only

will download the full soft file and thereafter extract info
}
\seealso{
Other GEO: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other methylation: 
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}

Other import: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{get.HPA.images}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseInhouseBeadStudio}()},
\code{\link{parseGeoSeriesMatrixSampleFile}()}
}
\concept{GEO}
\concept{genoset}
\concept{import}
\concept{methylation}
