% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\docType{data}
\name{palette.dl.annot}
\alias{palette.dl.annot}
\title{list of different colors used for plotting}
\format{
An object of class \code{list} of length 20.

An object of class \code{list} of length 20.
}
\usage{
palette.dl.annot

palette.dl.annot
}
\description{
\code{palette.dl.annot}
Object of type List. Vectors with different colors used for plotting

\code{palette.dl.annot}
Object of type List. Vectors with different colors used for plotting
}
\keyword{datasets}
