% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\docType{data}
\name{palette.dl.heatcol}
\alias{palette.dl.heatcol}
\title{list of different colors used for heatmaps
Object of type List. Vectors with different colors used for plotting}
\format{
An object of class \code{list} of length 15.

An object of class \code{list} of length 15.
}
\usage{
palette.dl.heatcol

palette.dl.heatcol
}
\description{
\code{palette.dl.heatcol}
Object of type List. Vectors with different colors used for plotting
}
\seealso{
Other plot: 
\code{\link{palette_gradients}}

Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{misc}
\concept{plot}
\keyword{datasets}
