% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{palette_gradientRamp}
\alias{palette_gradientRamp}
\title{Generate ramped color palettes - values and ramps described by vectors with llength n
This version (3) is different from palette_dl_heatmap2 since it uses palette numbers defined outside of function}
\usage{
palette_gradientRamp(
  x,
  palette.sat = c(-2, 2),
  my.pal = dlfoo2::palette_gradients[["blue2white2red"]],
  na.col = "gray",
  n.col = 50
)
}
\arguments{
\item{x}{a matrix containg expression values}

\item{palette.sat}{what saturation levels to use for palette. defaults to c(-2,2)}

\item{my.pal}{character vector of colors}

\item{na.col, }{color for NA-values. defaults to 'gray'}

\item{n.col, }{how menay steps in palette}

\item{pal.vec}{if supplying a palette and not use pre-defined}
}
\value{
a list object with 3 slots:  legend, col and val
}
\description{
Generate ramped color palettes - values and ramps described by vectors with llength n
This version (3) is different from palette_dl_heatmap2 since it uses palette numbers defined outside of function
}
\seealso{
Other plot functions: 
\code{\link{ggPlotEmpty}()},
\code{\link{heatmap.3}()}

Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{misc}
\concept{plot functions}
