% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\docType{data}
\name{palette_gradients}
\alias{palette_gradients}
\title{list of different colors used for heatmaps
Object of type List. Vectors with different colors used for plotting}
\format{
An object of class \code{list} of length 13.
}
\usage{
palette_gradients
}
\description{
list of different colors used for heatmaps
Object of type List. Vectors with different colors used for plotting
}
\seealso{
Other plot: 
\code{\link{palette.dl.heatcol}}

Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}
}
\concept{misc}
\concept{plot}
\keyword{datasets}
