% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_ImportExport.R
\name{parseGeoSeriesMatrixSampleFile}
\alias{parseGeoSeriesMatrixSampleFile}
\title{parse GEO series matrix to get sampel info sheet}
\usage{
parseGeoSeriesMatrixSampleFile(file.name = NULL)
}
\arguments{
\item{geo.id}{GEO GSE id}

\item{runName}{Fie dir-Name to use forthis run, e.g. rcc_me450.}

\item{runFolder}{Higher level directory to store processed genoset object}
}
\value{
save tables to disk
}
\description{
Wrapper for importing soft data from GEO using the GEOquery package

Methylation 450 data only

will download the full soft file and thereafter extract info
}
\seealso{
Other GEO: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseIDAT}()}

Other series matrix: 
\code{\link{geoSeriesMatrixPdataPolish}()}

Other genoset: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{gsSinkGenosetDescription}()},
\code{\link{liftoverGRanges}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meAnalysisCorTableStats}()},
\code{\link{meAnalysisCorrGexRandom}()},
\code{\link{meAnalysisCorrGex}()},
\code{\link{meFilterBetaNa}()},
\code{\link{meFilterVMF}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meHeatmapRCC}()},
\code{\link{meMergeCpGs}()},
\code{\link{meMergeGenoSetsBeta}()},
\code{\link{meNormalizeBeta}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}

Other import: 
\code{\link{geoSeriesMatrixPdataPolish}()},
\code{\link{get.HPA.images}()},
\code{\link{meGetGEOsoft_me450}()},
\code{\link{meParseBeadStudio}()},
\code{\link{meParseGDACv2}()},
\code{\link{meParseIDAT}()},
\code{\link{meParseInhouseBeadStudio}()}
}
\concept{GEO}
\concept{genoset}
\concept{import}
\concept{series matrix}
