% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{pca2df}
\alias{pca2df}
\title{produce a data frame from a PCA results object}
\usage{
pca2df(x)
}
\arguments{
\item{x}{results object from pca analysis (prcomp)}
}
\value{
data frame
}
\description{
Version 1.0 20190916
}
\seealso{
Other pca: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot}()}
}
\concept{pca}
