% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot colors}
\usage{
plot_colors(col.vec, text.vec = NULL)
}
\arguments{
\item{col.vec, }{vector with colors}

\item{text.vec, }{optional. define text to be plotted for each color}
}
\value{
a plot
}
\description{
\code{plot_colors}
Function to view different (hexadecimal) colors given as a character vector
}
