% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plot heatmap - genereic function}
\usage{
plotHeatmap(
  x,
  center.rows = T,
  cluster.rows = T,
  pdata = NULL,
  pdata.id.column = "sample_id",
  fdata = NULL,
  fdata.annot.columns = NULL,
  color.key.pdata,
  color.key.genes,
  pal.heatmap = rev(c(rev(RColorBrewer::brewer.pal(9, "YlOrRd")[1:9]), rep("#FFFFCC",
    1), rep("white", 1), rep("#FFFFCC", 1), RColorBrewer::brewer.pal(9, "YlGn")[1:9])),
  pal.sat = c(-3, 3),
  trace = "none",
  notecol = NA,
  dendrogram = "none",
  Rowv = NA,
  Colv = NA,
  labRow = NA,
  labCol = NA,
  ...
)
}
\arguments{
\item{x}{expression matrix. Input shpuld be Sorted and Filtered!}

\item{center.rows}{if to center rows}

\item{pdata}{data frame. sample colData with sample phenotype labels. Should contain 1 column with sample_id that match colnames for pdata/coldata. remaining columns will be plotted. The order of pdata will determine the sample-order to plot}

\item{pdata.id.column}{name of primary sample id column in pdata file. unique names used as colnames in matrix}

\item{fdata}{feature info if to plot gene-groups}

\item{fdata.annot.columns}{what fdata annotations to plot}

\item{color.key.pdata}{color key for annottaion of samples}

\item{color.key.genes}{color key for annotation of genes in signature}

\item{pal.heatmap}{What palette}

\item{pal.sat}{saturation levels for palette. detfaults to '-3,3'}

\item{runName}{name of plot/analyses}

\item{runFolder}{name of directory}

\item{amapDist}{Supplied to amap::Dist function. Can be ONE of following distance methods. "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.}
}
\value{
heatmaps to file
}
\description{
Version 1.0 20191018
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other heatmap: 
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{heatmapAnnotations}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{meHeatmapRCC}()}

Other gene expression: 
\code{\link{corrCcle}()},
\code{\link{corrGtex}()},
\code{\link{create.dummy.matrix}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetCreate}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{esetVarianceFilter}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaSurvPlot2}()}
}
\concept{gene expression}
\concept{heatmap}
\concept{plot wraps}
