% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{plotViolinScatterRankBarGeneric}
\alias{plotViolinScatterRankBarGeneric}
\title{plotViolinScatterRankBarGeneric}
\usage{
plotViolinScatterRankBarGeneric(
  x,
  gene.id = NULL,
  gene.id.2 = NULL,
  gene.id.signature = NULL,
  pdata = NULL,
  pdata.column = NULL,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  axis_names_n = T,
  rank.plot = F,
  stripchart.plot = F,
  violin.2nd.group.column = NULL,
  drop.na.annotation = T,
  barplot.y.freq = T,
  rank.decreasing = F,
  my.size = 2,
  my.alpha = 0.7,
  my.stroke = 0.05,
  my.stroke.color = "gray10",
  color.key = NULL,
  do.gradient = T,
  pseudo.gradient = T,
  gradient.sat = c(-1, 1),
  palette.gradient = rev(dlfoo2::palette_gradients[["red2white2blue"]]),
  gadient.na.col = NA,
  gradient.normalize = F,
  gradient.squish.probs = c(0.05, 0.95),
  my.samples = NULL,
  do.boxplot = F,
  swap.axes = F,
  showGuide = F,
  coord.fixed = F,
  xlim.cap = NULL,
  ylim.cap = NULL,
  xlim.rescale = NULL,
  ylim.rescale = NULL,
  zero.lines = F,
  shape.key = NULL,
  zoom.coord = NULL
)
}
\arguments{
\item{x}{results object. list (tSNEres) from run.tsne-analysis, or umap (umap object), or PCA (prcomp object). Can also be a data.frame (sample_id, dim1, dim2) or an expressionSet}

\item{gene.id}{supplied if x is an ExpressionSet and if scatter or rankplot}

\item{gene.id.2}{supplied if x is an ExpressionSet and if scatterplot}

\item{gene.id.signature}{supplied if x is expression set and if to create a score for violin plot over multiple genes}

\item{pdata}{Optional. Pairs 'sample_id' with the "annotation' used for group-colors. If not provided an 'annotation' column must be present in x. data frame with, minimally, columns 'sample_id' and a column for annotation (defined by 'annotation.column').}

\item{pdata.column}{should be a column in the pdata object to use for annotation (colors).}

\item{plot.title}{Title of plot}

\item{axis_names_n}{if to add number of samples to sample groups (x-axis names). Note: not feasible if downstream facet_wrap}

\item{rank.plot}{set to TRUE if to do Rank plot instead of Violin. Defaults to False. Possile if dim1 onöy is  supplied (not dim1 and dim2, then scatterplot).}

\item{stripchart.plot}{set to TRUE if to do stripchart instead of scatterplot. Provided as dim1 (values) and dim2 (groupings).}

\item{violin.2nd.group.column}{specify a columnname if to split violin groups into secondary groups}

\item{drop.na.annotation}{if to drop NA values from annotation column. defaults TRUE. May be set to FALSE for e.g. barplots to get a better representation of the data.}

\item{barplot.y.freq}{TRUE if barplot y-axis should be frequency percent, 0-100 (not total N)}

\item{rank.decreasing}{set to True if Rank plot and if rank decerasing on x-scale}

\item{my.size}{point size}

\item{my.alpha}{point alpha}

\item{my.stroke}{stroke width}

\item{my.stroke.color}{stroke line colors}

\item{color.key}{what color key to use for annotation. Detaults to dlfoo2::color_subtypes (but could be set to e.g. dlfoo2::color_annotations)}

\item{do.gradient}{set to FALSE if to manually override numeric check for gradient or manual fil}

\item{gradient.squish.probs}{if to cap squish gradient to remove effects on color scale of outliers. vector of tow percent values. Uses quantile-probs. Defaults to c(0.05, 0.95),i.e. cap scale at 5 and 95 of values}

\item{my.samples}{if to highligt samples - character vector of sample names}

\item{do.boxplot}{set to true if boxplot insead of violin}

\item{swap.axes}{if to swap x and y axes}

\item{showGuide}{if to show guide}

\item{coord.fixed}{if coord_fixed ggplot param}

\item{xlim.cap}{CAPs x - values (dim2) if any outsidde}

\item{ylim.cap}{CAPs x - values (dim2) if any outsidde}

\item{xlim.rescale}{passed on to axis. vector of 2. set upper/lower to NA if only rescaling one value}

\item{ylim.rescale}{passed on to axis. vector of 2. set upper/lower to NA if only rescaling one value}

\item{zero.lines}{if to higlight h- and -vlines through origo}

\item{shape.key}{if to plot annotations in different shapes (connected to annotation)}

\item{zoom.coord}{list of x and y coordinates to specify/zoom plot coordinates}

\item{color.key.gradient}{what gradient.}

\item{gradient.nomalize}{TRUE or FALSE if to z normalize annotation value (x-mean(x)) sd(x)}
}
\value{
a ggplot object
}
\description{
Generic plot for tSNE (list from tsneWrapper analysis), umap (umap object), or PCA (prcomp object) but also for generic data frames and for plotting Ranks only

If data.frama supplied. Then:

- Bar plot: Only when x is data.frame and both dim1 non-numeric characrer/factors. Primary annotation supplied in pdata/pdata.column.

- Violin: only dim1 numeric OR if eset, gene.id (but not gene.id_2) supplied

- Rank plot: as for Violin but rank.plot set to TRUE,

- Scatter plot: dim1 and dim2 numeric OR if eset, gene.id and gene.id.2 supplied
}
\seealso{
Other umap: 
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{umap2df}()}

Other tsne: 
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{tsne2df}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other pca: 
\code{\link{pca2df}()},
\code{\link{scatterPoint_annotationPlot}()}

Other ScatterPlot: 
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{limmaGeneCoOcPlot}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}
}
\concept{Barplot}
\concept{PlotWrap}
\concept{RankPlot}
\concept{ScatterPlot}
\concept{Violin}
\concept{expressonset}
\concept{pca}
\concept{tsne}
\concept{umap}
