% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{png_dl}
\alias{png_dl}
\title{wrapper for easy A4/A5 png plots}
\usage{
png_dl(file.name = NULL, Ax = "A4", orientation = "P")
}
\arguments{
\item{orientation, }{prortrait (P) or Landscape (L)}

\item{ax, what}{type of paper size (A4 or A5)}
}
\description{
\code{png_dl}
wrapper for easy A4/A5 png plots
}
