% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{rectMatrix}
\alias{rectMatrix}
\title{Small function to extract x-left and x-right values for when plotting rectangles, primarily using circlize package (segments within one sector specified by x-coordinates)}
\usage{
rectMatrix(x, x.offset = 0)
}
\arguments{
\item{x}{A vector vith integer values. Each value repesent the size of each reactangle.}

\item{x.offset}{if to adjust the x-value. Defaults to 0, which means rectangles start at x=0 and end at x=1 for input value 1}
}
\value{
Matrix with The cumulative sum, i.e. x-axis positions  (x-axis positions for x right and x left). A third column with mid-values is provided
}
\description{
Small function to extract x-left and x-right values for when plotting rectangles, primarily using circlize package (segments within one sector specified by x-coordinates)
}
\seealso{
Other plots: 
\code{\link{links2sectors}()}

Other circlize: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{links2sectors}()}
}
\concept{circlize}
\concept{circos.rect}
\concept{plots}
\concept{rect}
