% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{dend_rev_clusters}
\alias{dend_rev_clusters}
\title{dendextend package utils}
\usage{
dend_rev_clusters(dend, rev.clusters = c(T, F, F))
}
\arguments{
\item{dend, }{an object of class \code{\link[stats]{dendrogram}}}

\item{rev.clusters, }{vector of T or F values. Will define i) how many clusters (k) and ii) what subclusters that should be reversed.}
}
\value{
an object of class \code{\link[stats]{dendrogram}}
}
\description{
dendextend package utils
}
\section{Functions}{
\itemize{
\item \code{dend_rev_clusters}: the order of nodes within a specified cluster. Uses the \code{\link[dendextend]{rotate}}  and \code{\link[dlfoo]{dend_classify}} functions.
}}

\seealso{
Other hca: 
\code{\link{ConsensusClusterPlus.dl}()},
\code{\link{ConsensusCluster_process}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{hca}()}

Other transformation: 
\code{\link{cormat2indiv}()},
\code{\link{create.dummy.matrix}()},
\code{\link{dend_classify}()},
\code{\link{dend_reorder}()},
\code{\link{esetCenterGenes}()},
\code{\link{esetCombine}()},
\code{\link{esetGroupMeans}()},
\code{\link{esetMergeRows}()},
\code{\link{esetRankValues}()},
\code{\link{expandRanges}()},
\code{\link{hca}()},
\code{\link{liftoverGRanges}()},
\code{\link{meMergeCpGs}()},
\code{\link{meNormalizeBeta}()}
}
\concept{hca}
\concept{transformation}
