% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Shiny.R
\name{scatterPoint_annotationPlot_SHINY}
\alias{scatterPoint_annotationPlot_SHINY}
\title{scatterPoint_annotationPlot_SHINY}
\usage{
scatterPoint_annotationPlot_SHINY(
  results.object,
  pdata = NULL,
  default.annotation = "tax_simp",
  ...
)
}
\arguments{
\item{results.object}{results object. list (tSNEres) from run.tsne-analysis, or umap (umap object), or PCA (prcomp object)}

\item{...}{additional parameters as used in the scatterPoint_annotationPlot function}
}
\value{
a shiny app plot will open
}
\description{
Shiny app for plotting taxonomy/annotations on generic 2D scatter plot results objects, as PCA umap, tsne etc.
}
\seealso{
Other shiny: 
\code{\link{boxplot_SHINY_annotation}()},
\code{\link{shinyPlot_annotation}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other tsne: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{tsne2df}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other umap: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{umap2df}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}
}
\concept{PCA}
\concept{PlotWrap}
\concept{shiny}
\concept{tsne}
\concept{umap}
