% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_data_documentation.R
\docType{data}
\name{shape_subtypes}
\alias{shape_subtypes}
\title{Shepes key for TCGA subtpes}
\format{
Named character vector with pch shapes
}
\usage{
shape_subtypes
}
\description{
Shepes key for TCGA subtpes
}
\keyword{datasets}
