% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{shapeKeyFix}
\alias{shapeKeyFix}
\title{Match up a shape key to given a series of given unique annotation
Add shapes if annots are missing}
\usage{
shapeKeyFix(shape.key = NULL, annots = NULL, add.shape = "21")
}
\value{
shape_key vector where all annotatoions have a color
}
\description{
Match up a shape key to given a series of given unique annotation
Add shapes if annots are missing
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other color: 
\code{\link{colorKeyFix}()}
}
\concept{color}
\concept{plot wraps}
