% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Shiny.R
\name{shinyPlot_annotation}
\alias{shinyPlot_annotation}
\title{shinyPlot_annotation of two variables. Matrix provides Dim1 whereas dataframe matched with matrix colnames provides Dim2 and annotations..}
\usage{
shinyPlot_annotation(
  eset = NULL,
  mat = NULL,
  fdata = NULL,
  pdata = NULL,
  color.key = dlfoo2::color_subtypes,
  fdata.altName.column = NULL,
  default.xaxis.column = NULL,
  default.pdata.column = NULL,
  default.annot.column = NULL,
  x.lim = NULL,
  y.lim = NULL
)
}
\arguments{
\item{mat}{gene identifier. should}

\item{pdata}{sample data frame with sample annotations}

\item{color.key}{named character vector with sample bxp colors}

\item{fdata.altName.column}{alternative name column, e.g. hgnc_symbol}

\item{default.pdata.column}{name of preferred pdata annotations column (not needed)}

\item{...}{additional parameters}
}
\value{
a shiny app plot will open
}
\description{
Shiny app for plotting taxonomy/annotations on 2 generic variables, e.g. methylation vs purity

Matrix with matched fdata and pdata. matrix/fdata supplies Dim1, pdata dim2 and annotations

variation of the Shiny app for plotting taxonomy/annotations on generic 2D scatter plot results objects, as PCA umap, tsne etc.
}
\seealso{
Other shiny: 
\code{\link{boxplot_SHINY_annotation}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{tsneShinyIndivSampleEvaluation}()},
\code{\link{tsneShinyTaxEvaluation}()}
}
\concept{scatterplot}
\concept{scatterpoint}
\concept{shiny}
