% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{survPlot_pdata}
\alias{survPlot_pdata}
\title{Generic for plotting survival data from sample table separated on defined groups}
\usage{
survPlot_pdata(
  surv.tab,
  surv.bin.column = NULL,
  surv.time.column = NULL,
  annot.column = NULL,
  color.key = dlfoo2::color_subtypes,
  conf.int = TRUE,
  title.main = ""
)
}
\arguments{
\item{surv.bin.column}{colunmn name fopr surv bin}

\item{surv.time.column}{column in pdata for surv time}

\item{annot.column}{name of annotations to use}

\item{color.key}{what color key}

\item{conf.int}{if to add confindence intervals}

\item{pdata}{sample table}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
Generic for plotting survival data from sample table separated on defined groups
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other clinical: 
\code{\link{geneplotTcgaClinical}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}

Other survival: 
\code{\link{tcgaSurvPlot}()}
}
\concept{clinical}
\concept{kaplan}
\concept{plot wraps}
\concept{survival}
