% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Misc.R
\name{tcga_shortname}
\alias{tcga_shortname}
\title{Create short inhouse name for TCGA and TARGET samples}
\usage{
tcga_shortname(barcode, project, program = "TCGA")
}
\arguments{
\item{barcode, }{the TCGA aliquot barcode id, e.g. TCGA-KL-8323-01A-21R-2315-07}

\item{project, }{the TCGA project id, e.g. BRCA}

\item{program, }{TCGA or TARGET initiative}
}
\value{
character string of short barcode name
}
\description{
Shortnames for TCGA samples using barcode and project
}
\seealso{
Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{toChr}()}

Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}
}
\concept{misc}
\concept{tcga}
