% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Misc.R
\name{tcgaBarcodeSplitter}
\alias{tcgaBarcodeSplitter}
\title{splits a tcga barcode into its defined codes}
\usage{
tcgaBarcodeSplitter(x)
}
\arguments{
\item{x}{TCGA aliquot barcode id ('-' or '_' separated)}
}
\value{
barcode character vector of length 1
}
\description{
Version 1.0 20190108
}
\seealso{
Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}

Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()},
\code{\link{tcga_shortname}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}
}
\concept{misc}
\concept{tcga}
