% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Misc.R
\name{tcgaPdataPanCan}
\alias{tcgaPdataPanCan}
\title{Get individual sample pdata}
\usage{
tcgaPdataPanCan(aliquot.barcode, platform = NULL)
}
\arguments{
\item{aliquot.barcode, }{the TCGA aliquot barcode id, e.g. TCGA-KL-8323-01A-21R-2315-07}

\item{platform}{what platform from choices; "Genome_Wide_SNP_6","HumanMethylation27","HumanMethylation450","IlluminaGA_miRNASeq","IlluminaGA_RNASeqV2","IlluminaHiSeq_DNASeqC","IlluminaHiSeq_miRNASeq","IlluminaHiSeq_RNASeqV2","MDA_RPPA_Core")}
}
\value{
character string of short barcode name
}
\description{
Uses dlfoo2::pdata_tcgaPanCan pData table to get sample info

Incuding the sample_id (tcga shortname) for TCGA and TARGET samples: project + sample type + patient

This Id can be used to map samples between different platsforms - not sensitive to e.g. DNA/RNA type

Wrapper that uses tcga_shortname function as well as the PanCan curated sample lists

NOTE 1: Samples excluded by PanCan will be given no name

NOTE 2: Replicate id's may be generated here
}
\seealso{
Other misc: 
\code{\link{cormat2indiv}()},
\code{\link{fromChr}()},
\code{\link{genomicPosition}()},
\code{\link{palette.dl.heatcol}},
\code{\link{palette_gradientRamp}()},
\code{\link{palette_gradients}},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcga_shortname}()},
\code{\link{toChr}()}

Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()},
\code{\link{tcga_shortname}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}
}
\concept{misc}
\concept{tcga}
