% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{tcgaSurvPlot}
\alias{tcgaSurvPlot}
\title{plot clinical data TCGA PAN tissues}
\usage{
tcgaSurvPlot(surv_tab)
}
\arguments{
\item{surv_tab}{data frame (or tibble) with patient_barcode, fu_time, fu_status, and fu_group}
}
\value{
ggplot object
}
\description{
plot clinical data TCGA PAN tissues
}
\seealso{
Other plot wraps: 
\code{\link{VlnPlot_seurat_dl}()},
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()}

Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcga_shortname}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}

Other clinical: 
\code{\link{geneplotTcgaClinical}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()}

Other survival: 
\code{\link{survPlot_pdata}()}
}
\concept{clinical}
\concept{plot wraps}
\concept{survival}
\concept{tcga}
