% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{tsne2df}
\alias{tsne2df}
\title{produce a data frame from a tSNE results list}
\usage{
tsne2df(tSNEres)
}
\arguments{
\item{tSNEres}{results list (tSNEres) from run.tsne-analysis}
}
\value{
data frame
}
\description{
Version 1.0 20190213
}
\seealso{
Other tsne: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyTaxEvaluation}()}
}
\concept{tsne}
