% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_Shiny_indivSampleEvaluation.R
\name{tsneShinyIndivSampleEvaluation}
\alias{tsneShinyIndivSampleEvaluation}
\title{Start Shiny app for plotting taxonomy/annotations on tSNE analyses of GEX, Me etc. Highlights individual TCGA samples. Uses pre-defined data sets loaded from disk.}
\usage{
tsneShinyIndivSampleEvaluation(my_samples, id_type = "sample_id")
}
\arguments{
\item{id_type}{if sample id or barcode. defaults to dlfoo style sample_id}

\item{sample_id}{chracter string of TCGA sample or barcode ids}
}
\value{
a shiny app plot will open
}
\description{
Start Shiny app for plotting taxonomy/annotations on tSNE analyses of GEX, Me etc. Highlights individual TCGA samples. Uses pre-defined data sets loaded from disk.
}
\seealso{
Other tcga: 
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{tcgaBarcodeSplitter_v2}()},
\code{\link{tcgaBarcodeSplitter}()},
\code{\link{tcgaPdataPanCan}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()},
\code{\link{tcga_shortname}()}

Other shiny: 
\code{\link{boxplot_SHINY_annotation}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{shinyPlot_annotation}()},
\code{\link{tsneShinyTaxEvaluation}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyTaxEvaluation}()}
}
\concept{PlotWrap}
\concept{shiny}
\concept{tcga}
