% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dfoo2_ShinyReTaxEvaluation.R
\name{tsneShinyTaxEvaluation}
\alias{tsneShinyTaxEvaluation}
\title{Shiny app for plotting taxonomy/annotations on tSNE analyses of GEX, Me etc. Uses pre-defined data sets loaded from disk.}
\usage{
tsneShinyTaxEvaluation()
}
\value{
a shiny app plot will open
}
\description{
Shiny app for plotting taxonomy/annotations on tSNE analyses of GEX, Me etc. Uses pre-defined data sets loaded from disk.
}
\seealso{
Other tsne: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{tsne2df}()},
\code{\link{tsneGex}()},
\code{\link{tsneMe}()},
\code{\link{tsnePlot}()}

Other shiny: 
\code{\link{boxplot_SHINY_annotation}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{shinyPlot_annotation}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}

Other PlotWrap: 
\code{\link{chordPlot_ccHCA_to_reTax}()},
\code{\link{circosPlot_cc_hca_tree}()},
\code{\link{enrichmentCoocPlotrWrap}()},
\code{\link{fgseaCoocPlotrWrap}()},
\code{\link{genomeCircleSegments_legacy}()},
\code{\link{genomeCircleSegments}()},
\code{\link{heatmapAnnotations}()},
\code{\link{heatmapGeomRaster}()},
\code{\link{limmaGeneCoOcPlot}()},
\code{\link{meHeatmapRCC}()},
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()},
\code{\link{segmentPlotter}()},
\code{\link{tsnePlot}()},
\code{\link{tsneShinyIndivSampleEvaluation}()}
}
\concept{PlotWrap}
\concept{shiny}
\concept{tsne}
