% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_AnalysisWrap.R
\name{umap2df}
\alias{umap2df}
\title{produce a data frame from an umap results object}
\usage{
umap2df(x)
}
\arguments{
\item{x}{results object from umap analysis}
}
\value{
data frame
}
\description{
Version 1.0 20190916
}
\seealso{
Other umap: 
\code{\link{plotViolinScatterRankBarGeneric}()},
\code{\link{scatterPoint_annotationPlot_SHINY}()},
\code{\link{scatterPoint_annotationPlot}()}
}
\concept{umap}
