% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotWrap.R
\name{VlnPlot_seurat_dl}
\alias{VlnPlot_seurat_dl}
\alias{VlnPlot.dl}
\title{wrapper for VlnPlot in Seurat package}
\usage{
VlnPlot_seurat_dl(
  gene.vec,
  color.vec,
  x.text.size = 12,
  y.title.size = 12,
  ...
)
}
\arguments{
\item{gene.vec}{vector with gene symbols}

\item{orientation, }{prortrait (P) or Landscape (L)}
}
\value{
plot to quartz
}
\description{
wrapper for VlnPlot in Seurat package
}
\seealso{
Other plot wraps: 
\code{\link{colorKeyFix}()},
\code{\link{esetGexSignatures}()},
\code{\link{esetHeatmapRCC}()},
\code{\link{geneRankPlot}()},
\code{\link{geneplotCcle1156}()},
\code{\link{geneplotGtex}()},
\code{\link{geneplotKnepper}()},
\code{\link{geneplotPanTCGA_violin}()},
\code{\link{geneplotTcgaClinical}()},
\code{\link{geneplotTcgaPan}()},
\code{\link{geneplotTcgaRccTax}()},
\code{\link{me2gexHeatmapRCC}()},
\code{\link{plotHeatmap}()},
\code{\link{shapeKeyFix}()},
\code{\link{survPlot_pdata}()},
\code{\link{tcgaSurvPlot2}()},
\code{\link{tcgaSurvPlot}()}
}
\concept{plot wraps}
\concept{seurat}
