% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dlfoo2_PlotMisc.R
\name{widen.color.matrix}
\alias{widen.color.matrix}
\title{Widen a color matrix}
\usage{
widen.color.matrix(my.mat, x = 1)
}
\arguments{
\item{my.mat, }{a matrix}

\item{x, }{how many times each column shoud be widen}
}
\value{
a matrix
}
\description{
Widen a color matrix
}
